+++ src/config.ts
@@ -41,6 +41,7 @@ export type ConfigType = {
     maxFollowingAccountsToPull: number;
     mutexWarnSeconds: number;
     reloadFeaturesEveryNthOpen: number;
+    sleepBetweenCompletionMS: number;
     staleDataSeconds: StaleDataConfig;
     timeoutMS: number;
     staleDataTrendingSeconds: number;
@@ -122,6 +123,7 @@ export const Config: ConfigType = {
     mutexWarnSeconds: 5,                 // How long to wait before warning about a mutex lock
     numServersToCheck: 30,               // NUM_SERVERS_TO_CHECK
     reloadFeaturesEveryNthOpen: 9,       // RELOAD_FEATURES_EVERY_NTH_OPEN
+    sleepBetweenCompletionMS: 1_000,     // How long to wait between batches of Toot.completeToots() calls
     timeoutMS: 5_000,                    // Timeout for API calls
 
     // Trending tags and links