+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT RELEASE
 * Add `Config.isAppFilterVisible` boolean
 * Add `Toot.completedAt` and `shouldComplete()`, rename `Toot.setDependentProperties()` to `completeProperties()`
+* Add Config.sleepBetweenCompletionMS to avoid overloading browser with completeToots()
 
 ## v0.29.0
 * Defer processing of the most expensive Toot properties (trendingLinks, mostly, but other tag stuff too) until all data has arrived